ContactsCommon: Store all local contacts in only one account.
In the past, a second local account was introduced. This is used in many
more places than the original phone-local so change all references to
this local account to use the new one.

There are several places where exactly one local account is assumed,
such as in ContactProvider. We should combine all local contacts into
one account name and type.

Change-Id: I1781d009557ece05e9eaa078847756c428a8fbbd